requirements.txt
# Web Framework
#fastapi==0.104.1
#uvicorn[standard]==0.24.0
gunicorn==21.2.0
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Data Validation & Settings
#pydantic==2.5.0
#pydantic-settings==2.1.0
python-dotenv==1.0.0

# Data Processing
#pandas>=2.2.0
#numpy==1.26.2
#pyodbc==5.0.1
pyodbc>=4.0.30